% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_data}
\alias{img_data}
\title{Simulated imaging data for example}
\format{
A list of 11 data frames, each with 264 rows and 120 columns.
\describe{
Each row represents the time series for one region of interest in the brain,
a summary of multiple voxels.
}
}
\usage{
img_data
}
\description{
A dataset containing simulated imaging data, summarized
as one time series for each region of interest.
}
\keyword{datasets}
