% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_step1.R
\name{read.nifti.files}
\alias{read.nifti.files}
\title{STEP 1 IN PNC ANALYSIS - READ IN NIFTI FILE
Function summarizes imaging from nifti files into 264 regions with corresponding time series}
\usage{
read.nifti.files(
  file.names,
  path = NULL,
  method = "average",
  resolution = 2,
  template = NULL,
  PNC = FALSE
)
}
\arguments{
\item{file.names}{provide names of files}

\item{path}{optional path of files, needs to be included if path to files is not included in file.names}

\item{method}{method for summarizing within region. Default is averaging all voxels in a given region, "pca" also available if using the first prinicipal component.}

\item{resolution}{resolution of images. Template provided for 2mm and 3mm resolution, default is 2.}

\item{template}{option to provide an alternative template if not using standard 2mm or 3mm ones provided.}

\item{PNC}{option to specify is using PNC raw data, which needs to be converted to standard coordinates.}
}
\value{
List of files containing region level voxel summaries for each subject.
}
\description{
STEP 1 IN PNC ANALYSIS - READ IN NIFTI FILE
Function summarizes imaging from nifti files into 264 regions with corresponding time series
}
