% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{Simulated covariate data for example}
\format{A data frame with 11 rows (subjects) and 10 columns (covariates).
\describe{
  Each row contains the covariate values for one subject.
}}
\usage{
X
}
\description{
X matrix of covariates used to generate the connectivity matrix for an example.
}
\keyword{datasets}
