% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_step2.R
\name{network.estimate}
\alias{network.estimate}
\title{STEP 2 IN PNC ANALYSIS - CONVERT TO NETWORK
Obtain data in format time x 264 for each subject and convert to network.}
\usage{
network.estimate(images, lambda = FALSE)
}
\arguments{
\item{images}{list of images containing region level summaries of nifti data, one object per subject in the list.}

\item{lambda}{vector of length n (number of subjects) of lambda used in graphical lasso per subject .}
}
\value{
List of adjacency matrices for each subject.
}
\description{
STEP 2 IN PNC ANALYSIS - CONVERT TO NETWORK
Obtain data in format time x 264 for each subject and convert to network.
}
